PDFview add, minor tweeks to IE
setup and got PDFviewer working it mostly a copy
of IE, changed my desktop to a grid layout for auto
linning icons for when I have to make the site responsive
will be working on a notePad next with my contacts.

to-do
time and data on task bar
icons add to taskbar when openning windows
start menu where you can launch the apps too.