some refactoring, got date and time formated and show
did some clean up had to refactor my bring forward
to be more handled by my desktop so that
my task bar icons can bring my windows forward.
also got clicking out side of the menu to close the menu
or clicking on a window close it too now. and
clicking an icon in the menu closes it as well.
should be deployable now!